backend/requirements.txt
Django==5.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework-simplejwt==5.3.2
drf-yasg==1.21.7
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
whitenoise==6.6.0
